set up tab navigation using react navigation
