+++ nonCisMaleInCs/ios/Podfile
@@ -8,6 +8,8 @@ target 'nonCisMaleInCs' do
 
   use_react_native!(:path => config["reactNativePath"])
 
+  pod 'RNVectorIcons', :path => '../node_modules/react-native-vector-icons'
+
   target 'nonCisMaleInCsTests' do
     inherit! :complete
     # Pods for testing
